Added `fill_partially` and overload for `draw_vertices`
A new method `buffer_t:fill_partially` allows to specify the number of
bytes to copy to a buffer. Added as a separate method instead of
overloading `fill` because they have identical signatures.

`command_buffer_t:draw_vertices` now has a new overload that allow to
specify the number of vertices to draw.