Fixed `buffer_t::fill` with data size zero
This used to crash/fail an assertion, if the buffer to fill is on device
memory, due to trying to allocate a 0-byte staging buffer.
In the fixed version, both staging buffer creation and the copy command
are skipped if data size is zero, but the synchronization calls are
still executed, as they may be needed by the calling application.