Allow to modify renderpass create info (#56)
* Allow to modify renderpass create info

Added an optional parameter `aAlterCreateInfoBeforeCreation` to `create_renderpass()`.
This allows to modify the `RenderPassCreateInfo` immediately before the
renderpass is created.

* Unified aAlter*BeforeCreation functions

The `RenderPassCreateInfo` structure is now stored in a member, and
accessible via getter `create_info()`, similar to the rest of the
framework.
`aAlterConfigBeforeCreation` is now called _after_ the createInfo has
been constructed, and thus can directly modify it. This may be a
breaking change.
The additional `aAlterCreateInfoBeforeCreation` parameter has been
removed again.